% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloom.R
\name{bloom_contains}
\alias{bloom_contains}
\title{Query Items in a Bloom filter}
\usage{
bloom_contains(bl, x)
}
\arguments{
\item{bl}{the bloom filter created with bloom() or bloom_create()}

\item{x}{a string or character vector}
}
\description{
Check to see if an item exists in a previously created bloom filter object.
}
\details{
This is a wrapper around the bloom function.Returns a vector of logical values
}
\examples{
library(bloom)
bloomf<-bloom_create(capacity=2000, error_rate=0.05, filename='/tmp/bloom.bin')
t<-unlist(strsplit('this is a test',' '))
bloom_add(bloomf, t)
bloom_contains(bl,'this')
}
