% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloom.R
\name{bloom_create}
\alias{bloom_create}
\title{Create Bloom filter}
\usage{
bloom_create(
  capacity = 1000,
  error_rate = 0.05,
  filename = tempfile(fileext = ".bin"),
  exists = file.exists(filename)
)
}
\arguments{
\item{capacity}{the approximate expected capacity}

\item{error_rate}{the desired error rate}

\item{filename}{the location to store the filter}

\item{exists}{if the filter exists, load the new filter from a pre-existing file, otherwise
create a new filer.}
}
\description{
Create a new scaling, counting bloom filter, or load a bloom filter from a pre-existing file.
}
\details{
This is a wrapper around the bloom function.
}
\examples{
library(bloom)
bloomf<-bloom_create(capacity=2000, error_rate=0.05, filename='/tmp/bloom.bin')
}
