% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{branch}
\alias{branch}
\title{Create a data.table of all input combinations}
\usage{
branch(cols, retain_all)
}
\arguments{
\item{cols}{a data.table containing cols}

\item{retain_all}{hold back an `All` option for each col supplied in cols}
}
\description{
Create a data.table of all input combinations
}
