% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bloom.R
\name{bloom}
\alias{bloom}
\title{Bloom filter}
\usage{
bloom(capacity = 1000, error_rate = 0.05, filename = tempfile(fileext =
  ".bin"), exists = file.exists(filename))
}
\arguments{
\item{capacity}{the approximate expected capacity}

\item{error_rate}{the desired error rate}

\item{filename}{the location to store the filter}

\item{exists}{if the filter exists, load the new filter from a pre-existing file, otherwise
create a new filer.}
}
\description{
Create a new bloom filter, or load a bloom filter from a pre-existing file.
}
\examples{
library(bloom)
bloom <- bloom(capacity = 1000, error_rate = .05, filename = "/tmp/bloom.bin")
bloom$add("foo", 2)
bloom$contains("bar")
bloom$contains("foo")
bloom$remove("foo", 2)
bloom$contains("foo")
bloom$add("foo", 2)
rm(bloom)
bloom <- bloom(capacity = 1000, error_rate = .05, filename = "/tmp/bloom.bin", exists = TRUE)
bloom$contains("foo")
}

